no message
git-svn-id: https://ggnfs.svn.sourceforge.net/svnroot/ggnfs/trunk@150 80ef69ca-9341-0410-9d60-8acc7774fa31

+++ src/lasieve4/gnfs-lasieve4e.c
@@ -2476,7 +2476,7 @@ int lasieve()
 #if 0
 #define KLEINJUNG_MPQS
 #endif
-						  int nf;
+						  long nf;
                           mpz_t *mf;
 #ifdef KLEINJUNG_MPQS
